requirements.txt
cssmin==0.2.0
Django==1.10.2
django-assets==0.12
django-tastypie==0.13.3
django-widget-tweaks==1.4.1
eea.frame==0.5.2
jsmin==2.0.9
psycopg2==2.5.4
python-dateutil==2.5.3
python-mimeparse==1.5.5
requests==2.11.1
six==1.10.0
wsgiref==0.1.2
